+++ nfm-controller/src/reports/mod.rs
@@ -1,13 +1,13 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-mod prometheus_remote_write_proto;
 pub mod publisher;
+mod publisher_amp;
+mod publisher_amp_remote_write_proto;
 mod publisher_endpoint;
-mod publisher_prometheus_remote_write;
 pub mod report;
 pub mod report_otlp;
 
+pub use publisher_amp::ReportPublisherAmazonManagedPrometheus;
 pub use publisher_endpoint::ReportCompression;
-pub use publisher_prometheus_remote_write::ReportPublisherAmazonManagedPrometheus;
 pub use report::{CountersOverall, ProcessCounters};
