+++ nfm-controller/src/reports/publisher_amp_remote_write_proto.rs
